# docs/openapi.yaml
openapi: 3.0.0
info:
  title: MyTrello API
  version: 0.1.9
  contact:
    name: Shagalov Veaceslav
    url: t.me/slava_shagalov
servers:
  - url: http://localhost/api/v1
    description: Main server
...
paths:
...
  /lists/{id}/cards:
    parameters:
      - schema:
          type: integer
        name: id
        in: path
        required: true
    post:
      tags:
        - Lists
      summary: Create a new card
      description: Create a new card
      requestBody:
        description: A JSON object containing the data to create card.
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCard"
      responses:
        "200":
          description: A JSON object of created card.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Card"
        "400":
          $ref: "#/components/responses/ErrResponseBadRequest"
        "401":
          $ref: "#/components/responses/ErrUnauthorized"
        "500":
          $ref: "#/components/responses/ErrInternalServerError"
      security:
        - cookieAuth: [ ]
    get:
      tags:
        - Lists
      summary: Returns all cards by list id
      description: Returns all cards by list id
      responses:
        "200":
          description: A JSON with cards array
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cards"
        "401":
          $ref: "#/components/responses/ErrUnauthorized"
        "500":
          $ref: "#/components/responses/ErrInternalServerError"
      security:
        - cookieAuth: [ ]
  /cards:
    parameters:
      - name: title
        in: query
        description: Query string for card title
        required: false
        schema:
          type: string
      - name: content
        in: query
        description: Query string for card content
        required: false
        schema:
          type: string
    get:
      tags:
        - Cards
      summary: Returns user cards by filter parameters
      description: Returns user cards by filter parameters
      responses:
        "200":
          description: A JSON object with filtered user cards.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cards"
        "400":
          $ref: "#/components/responses/ErrResponseBadRequest"
        "401":
          $ref: "#/components/responses/ErrUnauthorized"
        "500":
          $ref: "#/components/responses/ErrInternalServerError"
      security:
        - cookieAuth: [ ]
  /cards/{id}:
    parameters:
      - schema:
          type: integer
        name: id
        in: path
        required: true
    get:
      tags:
        - Cards
      summary: Returns card by id
      description: Returns card by id
      responses:
        "200":
          description: A JSON object of card.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Card"
        "400":
          $ref: "#/components/responses/ErrResponseBadRequest"
        "401":
          $ref: "#/components/responses/ErrUnauthorized"
        "404":
          $ref: "#/components/responses/ErrCardNotFound"
        "500":
          $ref: "#/components/responses/ErrInternalServerError"
      security:
        - cookieAuth: [ ]
    patch:
      tags:
        - Cards
      summary: Partial update of card
      description: Partial update of card
      requestBody:
        description: A JSON object containing the data to update card.
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PartialUpdateCard"
      responses:
        "200":
          description: A JSON object of updated card.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Card"
        "400":
          $ref: "#/components/responses/ErrResponseBadRequest"
        "401":
          $ref: "#/components/responses/ErrUnauthorized"
        "403":
          $ref: "#/components/responses/ErrForbidden"
        "404":
          $ref: "#/components/responses/ErrCardNotFound"
      security:
        - cookieAuth: [ ]
    delete:
      tags:
        - Cards
      summary: Delete card
      description: Delete card
      responses:
        "204":
          description: Card deleted successfully.
        "400":
          $ref: "#/components/responses/ErrResponseBadRequest"
        "401":
          $ref: "#/components/responses/ErrUnauthorized"
        "404":
          $ref: "#/components/responses/ErrCardNotFound"
      security:
        - cookieAuth: [ ]
components:
  schemas:
    Id:
      type: integer
      required: true
      example: 3
...
    Time:
      type: string
      example: "11.07.2023 12:23:45"
...
    Card:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/Id"
        list_id:
          $ref: "#/components/schemas/Id"
        title:
          type: string
          example: "REST"
        content:
          type: string
          example: "RESTFUL API DOCS: https://restfulapi.net/"
        position:
          type: integer
          example: 5
        created_at:
          $ref: "#/components/schemas/Time"
        updated_at:
          $ref: "#/components/schemas/Time"
    Cards:
      type: object
      properties:
        cards:
          type: array
          items:
            $ref: "#/components/schemas/Card"
    CreateCard:
      type: object
      properties:
        title:
          type: string
          example: "REST"
        content:
          type: string
          example: "RESTFUL API DOCS: https://restfulapi.net/"
    PartialUpdateCard:
      type: object
      properties:
        list_id:
          required: false
          type: integer
          example: 3
        title:
          type: string
          required: false
          example: "REST"
        content:
          type: string
          required: false
          example: "RESTFUL API DOCS: https://restfulapi.net/"
        position:
          type: integer
          required: false
          example: 5
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: JSESSIONID
  responses:
    ErrResponseBadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            type: object
            description: Bad request
            properties:
              error:
                type: string
                example: "{{reason}}"
    ErrUnauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            type: object
            description: Unauthorized
            properties:
              error:
                type: string
                example: "unauthorized"
...
    ErrCardNotFound:
      description: Card not found
      content:
        application/json:
          schema:
            type: object
            description: Card not found
            properties:
              error:
                type: string
                example: "card not found"
    ErrForbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            type: object
            description: Access denied
            properties:
              error:
                type: string
                example: "access denied"
    ErrInternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            type: object
            description: Internal Server Error
            properties:
              error:
                type: string
                example: "{{reason}}"